+++ polybar-scripts/notification-xkcd/latest
@@ -1,0 +1,1 @@
+2369Read
