--- polybar-scripts/xkcd-notifier/latest
@@ -1,1 +1,0 @@
-2356Read
